rename assertion template to assertion
