+++ tests/testing/unit/mocks/middleware/intersection.tsx
@@ -1,8 +1,7 @@
 const { it } = intern.getInterface('bdd');
 const { describe } = intern.getPlugin('jsdom');
 import { tsx, create } from '../../../../../src/core/vdom';
-import renderer from '../../../../../src/testing/renderer';
-import assertionTemplate from '../../../../../src/testing/assertionTemplate';
+import renderer, { assertion } from '../../../../../src/testing/renderer';
 import createIntersectionMock from '../../../../../src/testing/mocks/middleware/intersection';
 import intersection from '../../../../../src/core/middleware/intersection';
 
@@ -15,10 +14,10 @@ describe('intersection mock', () => {
 			return <div key="root">{JSON.stringify(details)}</div>;
 		});
 		const r = renderer(() => <App key="app" />, { middleware: [[intersection, intersectionMock]] });
-		r.expect(assertionTemplate(() => <div key="root">{`{"intersectionRatio":0,"isIntersecting":false}`}</div>));
+		r.expect(assertion(() => <div key="root">{`{"intersectionRatio":0,"isIntersecting":false}`}</div>));
 		intersectionMock('root', { isIntersecting: true });
-		r.expect(assertionTemplate(() => <div key="root">{`{"isIntersecting":true}`}</div>));
+		r.expect(assertion(() => <div key="root">{`{"isIntersecting":true}`}</div>));
 		intersectionMock('root', { isIntersecting: false });
-		r.expect(assertionTemplate(() => <div key="root">{`{"isIntersecting":false}`}</div>));
+		r.expect(assertion(() => <div key="root">{`{"isIntersecting":false}`}</div>));
 	});
 });
